For day 9 I want a class Day9 in se.johannesdahlgren.aoc24 where a main method runs the following problem:
Read a file containing numbers 0-9 in a long sequence. The file is src/main/resources/day9

every other number contains the lenght of a block and the next contains the lenght of blank space between the block and the next block.

The result of this program should produce another sequence starting with index 0 and adding x of the index number where x is the length of the block and a . where there are blank space.

If there is uneven numbers in the input we should process the last number but without blank spaces at the end.

The language is Java with version 21 and I want to use all features of the 21 version