For day 11, I have a list of numbers, each turn, each number does one of the following, in the following order
if number is 0 - the new number is 1

if number has even number of digits, split it into two numbers where the first number is the first half of the original number and the other new number is the other half

If none of the above multiply the number by 2024

This program should run for 25 turns. How many numbers do I end up with after 25 turns. The initial list of numbers is located in a file src/main/resources/day11