For Day13 I need to know the cheapest way to win as many prices as possible.
Every A press cost 3 tokens and every B press costs 1 token, I need to be on the exact location to win a price. If I cant reach the price on a machine lets skip that machine.

In the input file src/main/resources/day13 we have several lines of input each machine takes 3 lines with a blank line between each machine.

For each machine the first line specifies how many places the a press moves in X and Y direction. The second  line specifies the B button and the last line specifies the location of the price.

Example

Button A: X+94, Y+34

Button B: X+22, Y+67

Prize: X=8400, Y=5400

How can I win a price with as few tokens as possible